# exp/orderbook: Speed up path finding dfs by avoiding calls to consumeOffers for visited nodes