While I was reviewing https://github.com/stellar/go/pull/3921 , I realized we could improve the performance of the path finding DFS by moving the visited check from the beginning of the function to the moment we consider each outgoing edge. 

Previously, we were doing **both** the call to `consumeOffers()` and the visited check for each outgoing edge. Now, we only call `consumeOffers()` if we pass the visited check. We improve the performance of the dfs by removing the expensive calls to `consumeOffers()` on edges which lead to nodes we have already visited.

Benchmark before changes:

```
goos: darwin
goarch: amd64
pkg: github.com/stellar/go/exp/orderbook
cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
BenchmarkMultipleDestinationAssets
BenchmarkMultipleDestinationAssets-12    	       2	 501200867 ns/op
BenchmarkTestData
BenchmarkTestData-12                     	       1	19009783799 ns/op
PASS
```

Benchmark after changes:

```
goos: darwin
goarch: amd64
pkg: github.com/stellar/go/exp/orderbook
cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
BenchmarkMultipleDestinationAssets
BenchmarkMultipleDestinationAssets-12    	       3	 402958734 ns/op
BenchmarkTestData
BenchmarkTestData-12                     	       1	15575703035 ns/op
PASS
```

There is about a 25% improvement in the benchmarks.